refactor(core): introduce generic PaymentProvider adapter layer
Replace hardcoded Stripe integration with a pluggable provider interface.
Core is now provider-agnostic, enabling future adapters (Polar, Paddle, etc.).

- Rename StripeRuntime → PaymentProvider interface with id/name fields
- Replace StripeProviderConfig with PayKitProviderConfig (id, name, createAdapter)
- Move Stripe SDK implementation (~900 lines) from core to @paykitjs/stripe
- Update all services: ctx.stripe.* → ctx.provider.*
- Remove stripe SDK dependency from core package
- Decouple CLI commands from Stripe-specific diagnostics
