+++ packages/paykit/package.json
@@ -68,7 +68,6 @@
     "pino": "^10.3.1",
     "pino-pretty": "^13.1.3",
     "posthog-node": "^5.28.8",
-    "stripe": "^19.1.0",
     "typescript": "^5.9.2",
     "zod": "^4.0.0"
   },
