+++ packages/stripe/package.json
@@ -29,7 +29,8 @@
     "typecheck": "tsc --build"
   },
   "dependencies": {
-    "paykitjs": "workspace:*"
+    "paykitjs": "workspace:*",
+    "stripe": "^19.1.0"
   },
   "devDependencies": {
     "tsdown": "^0.21.1",
