+++ packages/stripe/src/index.ts
@@ -1,3 +1,3 @@
 export { stripe } from "./stripe-provider";
 
-export type { StripeProviderOptions } from "./stripe-provider";
+export type { StripeOptions } from "./stripe-provider";
